requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
bs4==0.0.1
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dialogflow==0.5.2
docutils==0.14
Flask==1.0.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gensim==3.7.1
google-api-core==1.9.0
google-auth==1.6.3
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
lxml==4.3.3
MarkupSafe==1.1.1
nltk==3.4
numpy==1.16.2
pandas==0.24.2
paramiko==2.4.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyNaCl==1.3.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.1
soupsieve==1.9
SQLAlchemy==1.3.2
tornado==6.0.2
urllib3==1.24.1
Werkzeug==0.15.2
